[fix] fix the instruction OP_SLL

+++ src/cpu/execution/exe_inst_32.rs
@@ -96,7 +96,7 @@ pub fn exe_inst(inst: &Instruction, cpu: &mut CPU) {
         },
         OP_SLL    => {
             cpu.reg[inst.rd.unwrap() as usize] =
-                cpu.reg[inst.rs1.unwrap() as usize] << cpu.reg[inst.rs2.unwrap() as usize];
+                ((cpu.reg[inst.rs1.unwrap() as usize] as u32) << cpu.reg[inst.rs2.unwrap() as usize]) as i32;
         },
         OP_SLT    => {
             cpu.reg[inst.rd.unwrap() as usize] =
